Clean-up of the shutdown (including making sure that any dynamic memory is de-allocated).
Bug fixes in the serial writer queue to ensure that the slots are freed correctly when they are dequeued during a write action.

Bug fix for serial thread to ensure that it tests for termination while it's running.

Refactored the hardware registry so that modifier functions only impact variables at module-level.